Allow configuring the server from file
Using server.properties you can now configure the core functionality of
the server. It is still a simple document store without any real baked
in functionality for dynamic content, client certs etc. And probably
needs some testing to determine if / when an error occurs how exactly
it's forwarded back to the client