Perform basic directory listing
Personally, I would like this to be more configuratble. Perhaps a
template? That may be overkill for now. But at the very least have it
that when we're loading a directory listing we sort and prepend date
information to the title:

ISO 8601 present in the title
present in the file-name
creation timestamp
or off - if configured