Create multi-module setup
Break files into server and lib modules.
The server module will contain all of the server specific classes while
the lib module will contain common files that can be used for anyone who
would want to create something gemini related or a different
implementation of this server.

This is all still a WIP