Consumable logging from server
In order to make the logging more consumable I made it a TSV output:

Incoming messages:
[date]\tIN\t[url]\t[bytes]

Outgoing messages:
[date]\tOUT\t[status]\t[meta]\t[bytes]

This gives you most of the information you'll need. As for error logging
I haven't setup that appender yet, but being a Java app I am thinking
it's best to serve ther error.log as a standard JAVA output? with
stackstrace :shrug: