Mark package as deployable
You can now take the standalone-gemini-server and deploy the "output" in
some way. It's assembly'd into a tar.gz with all the necessary files you
may need. I am not sure the _best_ way to achieve deploying new versions
of it though? So I may need to think of something else?

Maybe your conf files can be mounted into the container? If that is the
case then you really just need to pull a docker image (if you so
choose).

If you're just running the server as-is. Then you would probably just
manually untar and copy your config into the dir or something