Normalize requested route before joinging docRoot
I makes more sense to just normalize first rather than join with the
docRoot then normalize, that just seems riskier. But you can compare the
previous commit with this to see two separate approaches.

This will take:

hostname/../index.gmi and normalize it to hostname/index.gmi

This behavior is common on the web so felt more natural