Log anonymized IPs
Returning a static IP instead to better comply with regulations and
privacy. This is something I've wanted to do for a long time. I decided
its better in the default implementation to log "something" where the IP
would go in a typical request log.

A feature request is to be able to configure this a bit better, as part
of the logging implementation, and have configurable "home IPs" to
bypass so you can separate external vs internal requests (to see any
non-self traffic or something)